@@ -14,8 +14,8 @@
     "@mui/icons-material": "5.5.1",
     "@mui/material": "5.5.3",
     "mapbox-gl": "2.7.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-router-dom": "6.2.2"
   },
   "devDependencies": {
